# performance improvelemts & cleanup