--- a/src/Yolov7net/YoloLabel.cs
+++ b/src/Yolov7net/YoloLabel.cs
@@ -5,8 +5,11 @@ namespace Yolov7net
     public class YoloLabel
     {
         public int Id { get; set; }
+
         public string? Name { get; set; }
+        
         public YoloLabelKind Kind { get; set; }
+        
         public Color Color { get; set; }
 
         public YoloLabel() => Color = Color.Yellow;