--- a/src/Yolov7net/YoloPrediction.cs
+++ b/src/Yolov7net/YoloPrediction.cs
@@ -5,7 +5,9 @@ namespace Yolov7net
     public class YoloPrediction
     {
         public YoloLabel? Label { get; set; }
+
         public RectangleF Rectangle { get; set; }
+        
         public float Score { get; set; }
 
         public YoloPrediction() { }