--- a/src/Yolov7net/Yolov5.cs
+++ b/src/Yolov7net/Yolov5.cs
@@ -13,24 +13,23 @@ namespace Yolov7net
     public class Yolov5 : IDisposable
     {
         private readonly InferenceSession _inferenceSession;
-        private YoloModel _model = new YoloModel();
+        private readonly YoloModel _model = new YoloModel();
 
-        public Yolov5(string ModelPath, bool useCuda = false)
+        public Yolov5(string modelPath, bool useCuda = false)
         {
 
             if (useCuda)
             {
                 SessionOptions opts = SessionOptions.MakeSessionOptionWithCudaProvider();
-                _inferenceSession = new InferenceSession(ModelPath, opts);
+                _inferenceSession = new InferenceSession(modelPath, opts);
             }
             else
             {
                 SessionOptions opts = new();
-                _inferenceSession = new InferenceSession(ModelPath, opts);
+                _inferenceSession = new InferenceSession(modelPath, opts);
             }
             
-            
-            /// Get model info
+            // Get model info
             get_input_details();
             get_output_details();
         }
@@ -56,17 +55,19 @@ namespace Yolov7net
                 _model.Confidence = conf_thres;
                 _model.MulConfidence = conf_thres + 0.05f;
             }
+
             if (iou_thres > 0f)
             {
                 _model.Overlap = iou_thres;
             }
-            return Supress(ParseOutput(Inference(image), image));
+
+            return Suppress(ParseOutput(Inference(image), image));
         }
 
         /// <summary>
         /// Removes overlaped duplicates (nms).
         /// </summary>
-        private List<YoloPrediction> Supress(List<YoloPrediction> items)
+        private List<YoloPrediction> Suppress(List<YoloPrediction> items)
         {
             var result = new List<YoloPrediction>(items);
 
@@ -78,7 +79,7 @@ namespace Yolov7net
 
                     var (rect1, rect2) = (item.Rectangle, current.Rectangle);
 
-                    RectangleF intersection = RectangleF.Intersect(rect1, rect2);
+                    var intersection = RectangleF.Intersect(rect1, rect2);
 
                     float intArea = intersection.Area(); // intersection area
                     float unionArea = rect1.Area() + rect2.Area() - intArea; // union area
@@ -99,7 +100,7 @@ namespace Yolov7net
 
         private DenseTensor<float>[] Inference(Image img)
         {
-            Bitmap resized = null;
+            Bitmap resized;
 
             if (img.Width != _model.Width || img.Height != _model.Height)
             {
@@ -115,14 +116,14 @@ namespace Yolov7net
                 NamedOnnxValue.CreateFromTensor("images", Utils.ExtractPixels2(resized))
             };
 
-            IDisposableReadOnlyCollection<DisposableNamedOnnxValue> result = _inferenceSession.Run(inputs); // run inference
+            var result = _inferenceSession.Run(inputs); // run inference
 
-            var output = new List<DenseTensor<float>>();
+            var output = new List<DenseTensor<float>>(_model.Outputs.Length);
 
             foreach (var item in _model.Outputs) // add outputs for processing
             {
                 output.Add(result.First(x => x.Name == item).Value as DenseTensor<float>);
-            };
+            }
 
             return output.ToArray();
         }
@@ -142,38 +143,38 @@ namespace Yolov7net
 
             var (xPad, yPad) = ((_model.Width - w * gain) / 2, (_model.Height - h * gain) / 2); // left, right pads
 
-            Parallel.For(0, (int)output.Length / _model.Dimensions, (i) =>
+            Parallel.For(0, (int)output.Length / _model.Dimensions, i =>
             {
-                if (output[0, i, 4] <= _model.Confidence) return; // skip low obj_conf results
+                var span = output.Buffer.Span.Slice(i * _model.Dimensions);
+                if (span[4] <= _model.Confidence) return; // skip low obj_conf results
 
-                Parallel.For(5, _model.Dimensions, (j) =>
+                for (int j = 5; j < _model.Dimensions; j++)
                 {
-                    output[0, i, j] = output[0, i, j] * output[0, i, 4]; // mul_conf = obj_conf * cls_conf
-                });
+                    span[j] *= span[4]; // mul_conf = obj_conf * cls_conf
+                }
 
-                Parallel.For(5, _model.Dimensions, (k) =>
+                float xMin = (span[0] - span[2] / 2 - xPad) / gain; // unpad bbox tlx to original
+                float yMin = (span[1] - span[3] / 2 - yPad) / gain; // unpad bbox tly to original
+                float xMax = (span[0] + span[2] / 2 - xPad) / gain; // unpad bbox brx to original
+                float yMax = (span[1] + span[3] / 2 - yPad) / gain; // unpad bbox bry to original
+
+                xMin = Utils.Clamp(xMin, 0, w - 0); // clip bbox tlx to boundaries
+                yMin = Utils.Clamp(yMin, 0, h - 0); // clip bbox tly to boundaries
+                xMax = Utils.Clamp(xMax, 0, w - 1); // clip bbox brx to boundaries
+                yMax = Utils.Clamp(yMax, 0, h - 1); // clip bbox bry to boundaries
+
+                for (int k = 5; k < _model.Dimensions; k++)
                 {
-                    if (output[0, i, k] <= _model.MulConfidence) return; // skip low mul_conf results
+                    if (span[k] <= _model.MulConfidence) continue; // skip low mul_conf results
 
-                    float xMin = ((output[0, i, 0] - output[0, i, 2] / 2) - xPad) / gain; // unpad bbox tlx to original
-                    float yMin = ((output[0, i, 1] - output[0, i, 3] / 2) - yPad) / gain; // unpad bbox tly to original
-                    float xMax = ((output[0, i, 0] + output[0, i, 2] / 2) - xPad) / gain; // unpad bbox brx to original
-                    float yMax = ((output[0, i, 1] + output[0, i, 3] / 2) - yPad) / gain; // unpad bbox bry to original
-
-                    xMin = Utils.Clamp(xMin, 0, w - 0); // clip bbox tlx to boundaries
-                    yMin = Utils.Clamp(yMin, 0, h - 0); // clip bbox tly to boundaries
-                    xMax = Utils.Clamp(xMax, 0, w - 1); // clip bbox brx to boundaries
-                    yMax = Utils.Clamp(yMax, 0, h - 1); // clip bbox bry to boundaries
-
-                    YoloLabel label = _model.Labels[k - 5];
-
-                    var prediction = new YoloPrediction(label, output[0, i, k])
+                    var label = _model.Labels[k - 5];
+                    var prediction = new YoloPrediction(label, span[k])
                     {
                         Rectangle = new RectangleF(xMin, yMin, xMax - xMin, yMax - yMin)
                     };
 
                     result.Add(prediction);
-                });
+                }
             });
 
             return result.ToList();
@@ -186,8 +187,7 @@ namespace Yolov7net
 
         private void prepare_input(Image img)
         {
-            Bitmap bmp = Utils.ResizeImage(img, _model.Width, _model.Height);
-
+            var bmp = Utils.ResizeImage(img, _model.Width, _model.Height);
         }
 
         private void get_input_details()