Output parsing is also chagned to use the tensor's Buffer instead of the indexer.
+Some cleanup.

Performance test (with the already existing Yolov7netPerformance project:

Before:
| Method |     Mean |    Error |   StdDev |   Median |     Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|------- |---------:|---------:|---------:|---------:|----------:|----------:|----------:|----------:|
| Yolov7 | 20.67 ms | 0.230 ms | 0.204 ms | 20.64 ms |  500.0000 |  500.0000 |  500.0000 |      5 MB |
| Yolov8 | 40.55 ms | 2.130 ms | 6.282 ms | 42.57 ms | 9000.0000 | 2000.0000 | 1000.0000 |    103 MB |
| Yolov5 | 20.58 ms | 0.388 ms | 0.399 ms | 20.51 ms |  750.0000 |  562.5000 |  562.5000 |      7 MB |

After:
|      Method |     Mean |    Error |    StdDev |     Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|------------ |---------:|---------:|----------:|----------:|----------:|----------:|----------:|
|      Yolov7 | 20.23 ms | 0.401 ms |  0.492 ms |  500.0000 |  500.0000 |  500.0000 |      5 MB |
|      Yolov8 | 16.67 ms | 0.333 ms |  0.342 ms |  500.0000 |  500.0000 |  500.0000 |      5 MB |
| Yolov8Numpy | 89.37 ms | 3.732 ms | 10.829 ms | 9000.0000 | 1000.0000 | 1000.0000 |    110 MB |
|      Yolov5 | 15.93 ms | 0.217 ms |  0.203 ms |  500.0000 |  500.0000 |  500.0000 |      5 MB |